LPS-62253 Move to right package
