+++ modules/apps/portlet-display-template/portlet-display-template-web/src/main/java/com/liferay/portlet/display/template/web/dynamic/data/mapping/util/PortletDisplayTemplateDDMPermissionSupport.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.display.template.web.ddm;
+package com.liferay.portlet.display.template.web.dynamic.data.mapping.util;
 
 import com.liferay.dynamic.data.mapping.util.DDMTemplatePermissionSupport;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
